debug add mul request iops test

+++ mul_request_iops.py
@@ -6,7 +6,8 @@ import time
 import multiprocessing
 import ray
 ray.init(address='auto', _node_ip_address='192.172.200.2')
-os.remove("record.txt")
+if os.path.exists("record.txt"):
+  os.remove("record.txt")
 #@ray.remote
 #def circle():
 #    return np.zeros(1000000)
